leetcode 977
I initially solved the problem using the built-in sort() function, but I wasn't satisfied with that approach. To gain more insight, I checked out Neetcode's video to understand how the question was explained and to identify any patterns or similarities in the problem. I didn’t watch the solution part; I stopped once I got the gist of the approach. Using that understanding, I then solved the problem independently.