leetcode 3042
Python makes life easy. I used startswith() and endswith(). First, I misinterpreted the question. I thought I just needed to check if the prefix of i and the prefix of j are equal, and the suffix of i and the suffix of j are equal, and then increase the count. But after many tries, I figured out that the word of i should be a prefix and suffix to the word of j.